create-diff-object: move addend math to a new function
Split out the addend offset math into a separate function so it can be
used elsewhere.
